- fix bug: not being able to parse more than 1 url - updated test cases to test for multiple url - code clean up

+++ helper/clean_db.py
@@ -5,8 +5,15 @@ import configs.search_engine_config as config
 
 
 def clean_test_db():
-    os.remove(config.TEST_DB_NAME)
+    if os.path.isfile(config.TEST_DB_NAME):
+        print("clean up test db ...")
+        os.remove(config.TEST_DB_NAME)
+    else:
+        pass
 
 
 def clean_dev_db():
-    os.remove(config.DEV_DB_NAME)
+    if os.path.isfile(config.DEV_DB_NAME):
+        os.remove(config.DEV_DB_NAME)
+    else:
+        pass
